requirements.txt
kafka-python
pandas
numpy
tensorflow
scikit-learn
joblib
flask
flask-socketio
eventlet
